+++ src/main/java/com/kalgory/kp/api/KalgoryKpApplication.java
@@ -1,0 +1,21 @@
+package com.kalgory.kp.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Spring Boot Application 실행.
+ */
+@SpringBootApplication
+public class KalgoryKpApplication {
+
+  /**
+   * entry point main function.
+   *
+   * @param args Process 인자.
+   */
+  public static void main(String[] args) {
+    SpringApplication.run(KalgoryKpApplication.class, args);
+  }
+
+}
