--- src/main/java/com/kalgory/kp/kalgorykp/KalgoryKpApplication.java
@@ -1,21 +1,0 @@
-package com.kalgory.kp.kalgorykp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Spring Boot Application 실행.
- */
-@SpringBootApplication
-public class KalgoryKpApplication {
-
-  /**
-   * entry point main function.
-   *
-   * @param args Process 인자.
-   */
-  public static void main(String[] args) {
-    SpringApplication.run(KalgoryKpApplication.class, args);
-  }
-
-}
