--- src/main/java/com/kalgory/kp/kalgorykp/config/SecurityConfiguration.java
@@ -1,49 +1,0 @@
-package com.kalgory.kp.kalgorykp.config;
-
-import java.util.List;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-/**
- * 스프링 시큐리티 설정 파일.
- */
-@Configuration
-@EnableWebSecurity
-@PropertySource("classpath:application-secure.properties")
-public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-
-  @Value("${HASH_ITERATIONS}")
-  private int HASH_ITERATIONS;
-
-  @Bean
-  public PasswordEncoder getPasswordEncoder() {
-    return new BCryptPasswordEncoder(HASH_ITERATIONS);
-  }
-
-  @Override
-  protected void configure(HttpSecurity http) throws Exception {
-    http
-        .formLogin().disable()
-        .cors().configurationSource(corsConfigurationSource()).and()
-        .csrf().disable();
-  }
-
-  private CorsConfigurationSource corsConfigurationSource() {
-    CorsConfiguration configuration = new CorsConfiguration();
-    configuration.addAllowedOriginPattern("*");
-    configuration.setAllowedHeaders(List.of("*"));
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    source.registerCorsConfiguration("/**", configuration);
-    return source;
-  }
-}
