+++ src/test/java/com/kalgory/kp/api/KalgoryKpApplicationTests.java
@@ -1,0 +1,15 @@
+package com.kalgory.kp.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.web.WebAppConfiguration;
+
+@WebAppConfiguration
+@SpringBootTest
+class KalgoryKpApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
