--- src/test/java/com/kalgory/kp/kalgorykp/KalgoryKpApplicationTests.java
@@ -1,15 +1,0 @@
-package com.kalgory.kp.kalgorykp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.web.WebAppConfiguration;
-
-@WebAppConfiguration
-@SpringBootTest
-class KalgoryKpApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
